refactor: Move config functions to common
