+++ components/dht.py
@@ -2,7 +2,7 @@ import sensors.dht as dht
 import functools
 import time
 import typing
-from config import SensorConfig
+from common import SensorConfig
 
 
 def run(config: SensorConfig, on_read: typing.Callable[[SensorConfig, dht.DHTReading], None]):
